Fold SuiteRunner into AbstractRunner
The distinction was rather arbitrary and only added to the confusion
between all the different things that partest calls “runner”.
