+++ src/partest/scala/tools/partest/nest/ConsoleRunner.scala
@@ -7,13 +7,8 @@ package scala.tools.partest
 package nest
 
 class ConsoleRunner(val config: RunnerSpec.Config) extends AbstractRunner {
-  val suiteRunner = new SuiteRunner (
-    testSourcePath = config.optSourcePath getOrElse PartestDefaults.sourcePath,
-    fileManager = new FileManager(ClassPath split PathResolver.Environment.javaUserClassPath map (Path(_))), // the script sets up our classpath for us via ant
-    updateCheck = config.optUpdateCheck,
-    failed = config.optFailed,
-    noexec = config.optNoExec,
-    nestUI = nestUI)
+  val testSourcePath = config.optSourcePath getOrElse PartestDefaults.sourcePath
+  val fileManager = new FileManager(ClassPath split PathResolver.Environment.javaUserClassPath map (Path(_))) // the script sets up our classpath for us via ant
 }
 
 object ConsoleRunner {
